+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 * `RSpec/InstanceVariable` - Recommend local variables in addition to `let`. ([@jaredbeck][])
+* Add `RSpec/ImplicitSubject` cop. ([@Darhazer][])
 
 ## 1.28.0 (2018-08-14)